Parallelize backup script
The two `for` loops could theoretically be replaced by one `find` command
piped to `parallel`, but the backup "names" need the `year` and `month`
values separately. Since I already have backups using those names, it's
better to keep the loops as-is.

Yes, I know, I could parse the year/date from the output of the `find`
command, but that feels like going too far in pursuit of a one-liner. I
want to be able to read this in a few years.