Fix backup failure handling
Adding 'set -e' to the exported function allows for failure if the
'duplicity' invocation fails.

Using --halt stops the process if 1 backup fails.

Normally, 'parallel' prints that a job failed at the *top* of the
output, but using `|| echo "BACKUP FAILED"` allows us to print that
message at the bottom. Using '&& exit 1' ensures that the process still
exits even though we "handled" the error with the echo command.